Move racetime to its own file
